Separate DoS classes from utxo whitelisting
